Fix narrowing warning

+++ ouzel/graphics/opengl/windows/RendererOGLWin.cpp
@@ -218,7 +218,7 @@ namespace ouzel
                     WGL_ALPHA_BITS_ARB, 8,
                     WGL_DEPTH_BITS_ARB, newDepth ? 24 : 0,
                     WGL_SAMPLE_BUFFERS_ARB, newSampleCount > 0 ? 1 : 0,
-                    WGL_SAMPLES_ARB, newSampleCount,
+                    WGL_SAMPLES_ARB, static_cast<int>(newSampleCount),
                     0,
                 };
